Enum hwloc::ObjectType

pub enum ObjectType {
    System,
    Machine,
    NUMANode,
    Package,
    Cache,
    Core,
    PU,
    Group,
    Misc,
    Bridge,
    PCIDevice,
    OSDevice,
    TypeMax,
}

Represents the type of a topology object.

Note that (partial) ordering for object types is implemented as a call into the hwloc library which defines ordering as follows:

A == B if ObjectType::A and ObjectType::B are the same.

A < B if ObjectType::A includes objects of type ObjectType::B.

A > B if objects of ObjectType::A are included in type ObjectType::B.

It can also help to think of it as comparing the relative depths of each type, so a ObjectType::System will be smaller than a ObjectType::PU since the system contains processing units.

Variants

System

The whole system that is accessible to hwloc. That may comprise several machines in SSI systems like Kerrighed.

Machine

The typical root object type. A set of processors and memory with cache coherency.

NUMANode

A set of processors around memory which the processors can directly access.

Package

Physical package, what goes into a socket. In the physical meaning, i.e. that you can add or remove physically.

Cache

The Cache. Can be L1i, L1d, L2, L3,...

Core

A computation unit (may be shared by several logical processors).

PU

Processing Unit, or (Logical) Processor.

An execution unit (may share a core with some other logical processors, e.g. in the case of an SMT core). Objects of this kind are always reported and can thus be used as fallback when others are not.

Group

Group objects.

Objects which do not fit in the above but are detected by hwloc and are useful to take into account for affinity. For instance, some operating systems expose their arbitrary processors aggregation this way. And hwloc may insert such objects to group NUMA nodes according to their distances.

These objects are ignored when they do not bring any structure.

Misc

Miscellaneous objects.

Objects without particular meaning, that can e.g. be added by the application for its own use, or by hwloc for miscellaneous objects such as MemoryModule (DIMMs).

Bridge

Any bridge that connects the host or an I/O bus, to another I/O bus.

Bridge objects have neither CPU sets nor node sets. They are not added to the topology unless I/O discovery is enabled through the custom flags.

PCIDevice

PCI device.

These objects have neither CPU sets nor node sets. They are not added to the topology unless I/O discovery is enabled through the custom flags.

OSDevice

Operating system device.

These objects have neither CPU sets nor node sets. They are not added to the topology unless I/O discovery is enabled through the custom flags.

TypeMax

An internal sentinel value.