Enum hwloc::TypeDepthError

pub enum TypeDepthError {
    TypeDepthUnknown,
    TypeDepthMultiple,
    TypeDepthBridge,
    TypeDepthPCIDevice,
    TypeDepthOSDevice,
    UnkownTypeDepthError,
}

Variants

TypeDepthUnknown

No object of given type exists in the topology.

TypeDepthMultiple

Objects of given type exist at different depth in the topology.

TypeDepthBridge

Virtual depth for bridge object level.

TypeDepthPCIDevice

Virtual depth for PCI device object level.

TypeDepthOSDevice

Virtual depth for software device object level.

UnkownTypeDepthError

HWLOC returned a depth error which is not known to the rust binding.