Struct hwloc::TopologyMemBindSupport

pub struct TopologyMemBindSupport {
    // some fields omitted
}

Flags describing actual memory binding support for this topology.

Methods

impl TopologyMemBindSupport

fn set_current_process(&self) -> bool

Binding the whole current process is supported.

fn get_current_process(&self) -> bool

Getting the binding of the whole current process is supported.

fn set_process(&self) -> bool

Binding a whole given process is supported.

fn get_process(&self) -> bool

Getting the binding of a whole given process is supported.

fn set_current_thread(&self) -> bool

Binding the current thread only is supported.

fn get_current_thread(&self) -> bool

Getting the binding of the current thread only is supported.

fn set_area(&self) -> bool

Binding a given memory area is supported.

fn get_area(&self) -> bool

Getting the binding of a given memory area is supported.

fn alloc(&self) -> bool

Allocating a bound memory area is supported.

fn first_touch(&self) -> bool

First-touch policy is supported.

fn bind(&self) -> bool

Bind policy is supported.

fn interleave(&self) -> bool

Interleave policy is supported.

fn replicate(&self) -> bool

Replication policy is supported.

fn next_touch(&self) -> bool

Next-touch migration policy is supported.

fn migrate(&self) -> bool

Migration flags is supported.